Email templates Updating template names and parameters.
