+++ src/services/emails/reset-password.js
@@ -2,17 +2,16 @@ const Promise = require('bluebird')
 const env = require('../../config/env')
 const transporter = require('../../config/nodemailer')
 
-module.exports = (req, resetToken) => {
+module.exports = user => {
   let template = {
     from: env('EMAIL'),
-    to: req.user.email,
-    subject: 'Style Point: Reset your password',
+    to: user.email,
+    subject: 'Style Point: Password Changed',
     html: `
         <p>
-            You are receiving this because you (or someone else) have requested the reset of the password for your account.
-            <br>Please click on the following link, or paste this into your browser to complete the process:<br><br>
-            http://${req.headers.host}/reset-password/${resetToken}<br><br>
-            If you did not request this, please ignore this email and your password will remain unchanged.
+            You are receiving this email because you changed your password.
+            <br>
+            If you did not request this change, please contact us immediately.          
         </p>
     `
   }
